move limit stream to live with limit job
This moves the limit stream out of the streaming package since its
behavior is specific to the limit job. I moved it to colocate it with
the limit job and moved hte limit job into its own file.
