+++ internal/search/streaming/stream.go
@@ -1,7 +1,6 @@
 package streaming
 
 import (
-	"context"
 	"sync"
 	"time"
 
@@ -17,63 +16,6 @@ type SearchEvent struct {
 
 type Sender interface {
 	Send(SearchEvent)
-}
-
-type LimitStream struct {
-	s         Sender
-	cancel    context.CancelFunc
-	remaining atomic.Int64
-}
-
-func (s *LimitStream) Send(event SearchEvent) {
-	count := int64(event.Results.ResultCount())
-
-	// Avoid limit checks if no change to result count.
-	if count == 0 {
-		s.s.Send(event)
-		return
-	}
-
-	// Get the remaining count before and after sending this event
-	after := s.remaining.Sub(count)
-	before := after + count
-
-	// Check if the event needs truncating before being sent
-	if after < 0 {
-		limit := before
-		if before < 0 {
-			limit = 0
-		}
-		event.Results.Limit(int(limit))
-	}
-
-	// Send the maybe-truncated event. We want to always send the event
-	// even if we truncate it to zero results in case it has stats on it
-	// that we care about it.
-	s.s.Send(event)
-
-	// Send the IsLimitHit event and call cancel exactly once. This will
-	// only trigger when the result count of an event causes us to cross
-	// the zero-remaining threshold.
-	if after <= 0 && before > 0 {
-		s.s.Send(SearchEvent{Stats: Stats{IsLimitHit: true}})
-		s.cancel()
-	}
-}
-
-// WithLimit returns a child Stream of parent as well as a child Context of
-// ctx. The child stream passes on all events to parent. Once more than limit
-// ResultCount are sent on the child stream the context is canceled and an
-// IsLimitHit event is sent.
-//
-// Canceling this context releases resources associated with it, so code
-// should call cancel as soon as the operations running in this Context and
-// Stream are complete.
-func WithLimit(ctx context.Context, parent Sender, limit int) (context.Context, Sender, context.CancelFunc) {
-	ctx, cancel := context.WithCancel(ctx)
-	stream := &LimitStream{cancel: cancel, s: parent}
-	stream.remaining.Store(int64(limit))
-	return ctx, stream, cancel
 }
 
 type StreamFunc func(SearchEvent)
